Implemented majority of opcodes

+++ src/ca/plasticglass/chip8/Memory.java
@@ -6,27 +6,27 @@ package ca.plasticglass.chip8;
  */
 public class Memory {
 
-    private static byte[] memory;
-    private final byte PROGRAM_LOAD_POINT = (byte) 0x200;
+    private static short[] memory;
+    private final short PROGRAM_LOAD_POINT = (short) 0x200;
     private final short MEMORY_SIZE = 0xFFF + 1; //+1 because of zero index
 
     public Memory(){
-        memory = new byte[MEMORY_SIZE];
+        memory = new short[MEMORY_SIZE];
     }
 
     public void loadFile(String fileName){
 
     }
 
-    public byte[] getMemory(){
+    public short[] getMemory(){
         return memory;
     }
 
-    public byte get(int address){
+    public short get(int address){
         return memory[address];
     }
 
-    public byte getLoadPoint(){
+    public short getLoadPoint(){
         return PROGRAM_LOAD_POINT;
     }
